FormItem: Pass full form model to form item validation
In some cases a form field needs complex validation logic that depends on other fields in the form (for example, a form has a country select and a phone number / bank card input that is validated by selected country). So the custom `validator` function needs access to the full form model.
The `descriptor` object contains only rules for the current field anyway, so extra fields on the validated model won't affect the validation.