message-box: close/cancel
We needed 3 states for our confirm. In Tipe if you make an edit and leave the page we ask if they want to Cancel (discard changes) or Success (save and continue). We needed a Close state for when they click on {x] or click out of the modal to close the modal without doing anything
* Success
* Cancel
* Close