Dialog: fixed the timing whne dialog content is recreated (#17124)
Changed the timing of the key updating to when the dialog is displayed. This will cause the component recreation happening while the dialog is being displayed, not while it is being closed.

`this.$nextTick()` should be avoided. Otherwise, vue will render the component once, and then the key is updated, the component will be rendered again.